LPS-120036 Move all the portal stuff to the SPI package.
