+++ modules/apps/change-tracking/change-tracking-service/src/main/java/com/liferay/change/tracking/internal/spi/reference/ExpandoValueTableReferenceDefinition.java
@@ -12,14 +12,18 @@
  * details.
  */
 
-package com.liferay.change.tracking.internal.reference.portal;
+package com.liferay.change.tracking.internal.spi.reference;
 
 import com.liferay.change.tracking.spi.reference.TableReferenceDefinition;
 import com.liferay.change.tracking.spi.reference.builder.ChildTableReferenceInfoBuilder;
 import com.liferay.change.tracking.spi.reference.builder.ParentTableReferenceInfoBuilder;
-import com.liferay.portal.kernel.model.CountryTable;
+import com.liferay.expando.kernel.model.ExpandoColumnTable;
+import com.liferay.expando.kernel.model.ExpandoRowTable;
+import com.liferay.expando.kernel.model.ExpandoTableTable;
+import com.liferay.expando.kernel.model.ExpandoValueTable;
+import com.liferay.expando.kernel.service.persistence.ExpandoValuePersistence;
+import com.liferay.portal.kernel.model.CompanyTable;
 import com.liferay.portal.kernel.service.persistence.BasePersistence;
-import com.liferay.portal.kernel.service.persistence.CountryPersistence;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -28,32 +32,45 @@ import org.osgi.service.component.annotations.Reference;
  * @author Preston Crary
  */
 @Component(service = TableReferenceDefinition.class)
-public class CountryTableReferenceDefinition
-	implements TableReferenceDefinition<CountryTable> {
+public class ExpandoValueTableReferenceDefinition
+	implements TableReferenceDefinition<ExpandoValueTable> {
 
 	@Override
 	public void defineChildTableReferences(
-		ChildTableReferenceInfoBuilder<CountryTable>
+		ChildTableReferenceInfoBuilder<ExpandoValueTable>
 			childTableReferenceInfoBuilder) {
 	}
 
 	@Override
 	public void defineParentTableReferences(
-		ParentTableReferenceInfoBuilder<CountryTable>
+		ParentTableReferenceInfoBuilder<ExpandoValueTable>
 			parentTableReferenceInfoBuilder) {
+
+		parentTableReferenceInfoBuilder.singleColumnReference(
+			ExpandoValueTable.INSTANCE.companyId,
+			CompanyTable.INSTANCE.companyId
+		).singleColumnReference(
+			ExpandoValueTable.INSTANCE.tableId,
+			ExpandoTableTable.INSTANCE.tableId
+		).singleColumnReference(
+			ExpandoValueTable.INSTANCE.columnId,
+			ExpandoColumnTable.INSTANCE.columnId
+		).singleColumnReference(
+			ExpandoValueTable.INSTANCE.rowId, ExpandoRowTable.INSTANCE.rowId
+		);
 	}
 
 	@Override
 	public BasePersistence<?> getBasePersistence() {
-		return _countryPersistence;
+		return _expandoValuePersistence;
 	}
 
 	@Override
-	public CountryTable getTable() {
-		return CountryTable.INSTANCE;
+	public ExpandoValueTable getTable() {
+		return ExpandoValueTable.INSTANCE;
 	}
 
 	@Reference
-	private CountryPersistence _countryPersistence;
+	private ExpandoValuePersistence _expandoValuePersistence;
 
 }
