+++ modules/apps/change-tracking/change-tracking-service/src/main/java/com/liferay/change/tracking/internal/spi/reference/UserGroupGroupRoleTableReferenceDefinition.java
@@ -12,15 +12,18 @@
  * details.
  */
 
-package com.liferay.change.tracking.internal.reference.portal;
+package com.liferay.change.tracking.internal.spi.reference;
 
 import com.liferay.change.tracking.spi.reference.TableReferenceDefinition;
 import com.liferay.change.tracking.spi.reference.builder.ChildTableReferenceInfoBuilder;
 import com.liferay.change.tracking.spi.reference.builder.ParentTableReferenceInfoBuilder;
 import com.liferay.portal.kernel.model.CompanyTable;
-import com.liferay.portal.kernel.model.ImageTable;
+import com.liferay.portal.kernel.model.GroupTable;
+import com.liferay.portal.kernel.model.RoleTable;
+import com.liferay.portal.kernel.model.UserGroupGroupRoleTable;
+import com.liferay.portal.kernel.model.UserGroupTable;
 import com.liferay.portal.kernel.service.persistence.BasePersistence;
-import com.liferay.portal.kernel.service.persistence.ImagePersistence;
+import com.liferay.portal.kernel.service.persistence.UserGroupGroupRolePersistence;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -29,35 +32,45 @@ import org.osgi.service.component.annotations.Reference;
  * @author Preston Crary
  */
 @Component(service = TableReferenceDefinition.class)
-public class ImageTableReferenceDefinition
-	implements TableReferenceDefinition<ImageTable> {
+public class UserGroupGroupRoleTableReferenceDefinition
+	implements TableReferenceDefinition<UserGroupGroupRoleTable> {
 
 	@Override
 	public void defineChildTableReferences(
-		ChildTableReferenceInfoBuilder<ImageTable>
+		ChildTableReferenceInfoBuilder<UserGroupGroupRoleTable>
 			childTableReferenceInfoBuilder) {
 	}
 
 	@Override
 	public void defineParentTableReferences(
-		ParentTableReferenceInfoBuilder<ImageTable>
+		ParentTableReferenceInfoBuilder<UserGroupGroupRoleTable>
 			parentTableReferenceInfoBuilder) {
 
 		parentTableReferenceInfoBuilder.singleColumnReference(
-			ImageTable.INSTANCE.companyId, CompanyTable.INSTANCE.companyId);
+			UserGroupGroupRoleTable.INSTANCE.companyId,
+			CompanyTable.INSTANCE.companyId
+		).singleColumnReference(
+			UserGroupGroupRoleTable.INSTANCE.userGroupId,
+			UserGroupTable.INSTANCE.userGroupId
+		).singleColumnReference(
+			UserGroupGroupRoleTable.INSTANCE.groupId,
+			GroupTable.INSTANCE.groupId
+		).singleColumnReference(
+			UserGroupGroupRoleTable.INSTANCE.roleId, RoleTable.INSTANCE.roleId
+		);
 	}
 
 	@Override
 	public BasePersistence<?> getBasePersistence() {
-		return _imagePersistence;
+		return _userGroupGroupRolePersistence;
 	}
 
 	@Override
-	public ImageTable getTable() {
-		return ImageTable.INSTANCE;
+	public UserGroupGroupRoleTable getTable() {
+		return UserGroupGroupRoleTable.INSTANCE;
 	}
 
 	@Reference
-	private ImagePersistence _imagePersistence;
+	private UserGroupGroupRolePersistence _userGroupGroupRolePersistence;
 
 }
