+++ modules/apps/change-tracking/change-tracking-service/src/main/java/com/liferay/change/tracking/internal/spi/reference/UserGroupTableReferenceDefinition.java
@@ -12,16 +12,17 @@
  * details.
  */
 
-package com.liferay.change.tracking.internal.reference.portal;
+package com.liferay.change.tracking.internal.spi.reference;
 
 import com.liferay.change.tracking.spi.reference.TableReferenceDefinition;
 import com.liferay.change.tracking.spi.reference.builder.ChildTableReferenceInfoBuilder;
 import com.liferay.change.tracking.spi.reference.builder.ParentTableReferenceInfoBuilder;
-import com.liferay.expando.kernel.model.ExpandoColumnTable;
-import com.liferay.expando.kernel.model.ExpandoTableTable;
-import com.liferay.expando.kernel.service.persistence.ExpandoColumnPersistence;
 import com.liferay.portal.kernel.model.CompanyTable;
+import com.liferay.portal.kernel.model.GroupTable;
+import com.liferay.portal.kernel.model.UserGroup;
+import com.liferay.portal.kernel.model.UserGroupTable;
 import com.liferay.portal.kernel.service.persistence.BasePersistence;
+import com.liferay.portal.kernel.service.persistence.UserGroupPersistence;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -30,40 +31,48 @@ import org.osgi.service.component.annotations.Reference;
  * @author Preston Crary
  */
 @Component(service = TableReferenceDefinition.class)
-public class ExpandoColumnTableReferenceDefinition
-	implements TableReferenceDefinition<ExpandoColumnTable> {
+public class UserGroupTableReferenceDefinition
+	implements TableReferenceDefinition<UserGroupTable> {
 
 	@Override
 	public void defineChildTableReferences(
-		ChildTableReferenceInfoBuilder<ExpandoColumnTable>
+		ChildTableReferenceInfoBuilder<UserGroupTable>
 			childTableReferenceInfoBuilder) {
+
+		childTableReferenceInfoBuilder.classNameReference(
+			UserGroupTable.INSTANCE.userGroupId, GroupTable.INSTANCE.classPK,
+			UserGroup.class
+		).resourcePermissionReference(
+			UserGroupTable.INSTANCE.userGroupId, UserGroup.class
+		).systemEventReference(
+			UserGroupTable.INSTANCE.userGroupId, UserGroup.class
+		);
 	}
 
 	@Override
 	public void defineParentTableReferences(
-		ParentTableReferenceInfoBuilder<ExpandoColumnTable>
+		ParentTableReferenceInfoBuilder<UserGroupTable>
 			parentTableReferenceInfoBuilder) {
 
 		parentTableReferenceInfoBuilder.singleColumnReference(
-			ExpandoColumnTable.INSTANCE.companyId,
-			CompanyTable.INSTANCE.companyId
-		).singleColumnReference(
-			ExpandoColumnTable.INSTANCE.tableId,
-			ExpandoTableTable.INSTANCE.tableId
+			UserGroupTable.INSTANCE.companyId, CompanyTable.INSTANCE.companyId
+		).parentColumnReference(
+			UserGroupTable.INSTANCE.userGroupId,
+			UserGroupTable.INSTANCE.parentUserGroupId
 		);
 	}
 
 	@Override
 	public BasePersistence<?> getBasePersistence() {
-		return _expandoColumnPersistence;
+		return _userGroupPersistence;
 	}
 
 	@Override
-	public ExpandoColumnTable getTable() {
-		return ExpandoColumnTable.INSTANCE;
+	public UserGroupTable getTable() {
+		return UserGroupTable.INSTANCE;
 	}
 
 	@Reference
-	private ExpandoColumnPersistence _expandoColumnPersistence;
+	private UserGroupPersistence _userGroupPersistence;
 
 }
