+++ modules/apps/change-tracking/change-tracking-service/src/main/java/com/liferay/change/tracking/internal/spi/reference/VirtualHostTableReferenceDefinition.java
@@ -12,17 +12,16 @@
  * details.
  */
 
-package com.liferay.change.tracking.internal.reference.portal;
+package com.liferay.change.tracking.internal.spi.reference;
 
 import com.liferay.change.tracking.spi.reference.TableReferenceDefinition;
 import com.liferay.change.tracking.spi.reference.builder.ChildTableReferenceInfoBuilder;
 import com.liferay.change.tracking.spi.reference.builder.ParentTableReferenceInfoBuilder;
 import com.liferay.portal.kernel.model.CompanyTable;
-import com.liferay.portal.kernel.model.GroupTable;
-import com.liferay.portal.kernel.model.UserGroup;
-import com.liferay.portal.kernel.model.UserGroupTable;
+import com.liferay.portal.kernel.model.LayoutSetTable;
+import com.liferay.portal.kernel.model.VirtualHostTable;
 import com.liferay.portal.kernel.service.persistence.BasePersistence;
-import com.liferay.portal.kernel.service.persistence.UserGroupPersistence;
+import com.liferay.portal.kernel.service.persistence.VirtualHostPersistence;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -31,48 +30,39 @@ import org.osgi.service.component.annotations.Reference;
  * @author Preston Crary
  */
 @Component(service = TableReferenceDefinition.class)
-public class UserGroupTableReferenceDefinition
-	implements TableReferenceDefinition<UserGroupTable> {
+public class VirtualHostTableReferenceDefinition
+	implements TableReferenceDefinition<VirtualHostTable> {
 
 	@Override
 	public void defineChildTableReferences(
-		ChildTableReferenceInfoBuilder<UserGroupTable>
+		ChildTableReferenceInfoBuilder<VirtualHostTable>
 			childTableReferenceInfoBuilder) {
-
-		childTableReferenceInfoBuilder.classNameReference(
-			UserGroupTable.INSTANCE.userGroupId, GroupTable.INSTANCE.classPK,
-			UserGroup.class
-		).resourcePermissionReference(
-			UserGroupTable.INSTANCE.userGroupId, UserGroup.class
-		).systemEventReference(
-			UserGroupTable.INSTANCE.userGroupId, UserGroup.class
-		);
 	}
 
 	@Override
 	public void defineParentTableReferences(
-		ParentTableReferenceInfoBuilder<UserGroupTable>
+		ParentTableReferenceInfoBuilder<VirtualHostTable>
 			parentTableReferenceInfoBuilder) {
 
 		parentTableReferenceInfoBuilder.singleColumnReference(
-			UserGroupTable.INSTANCE.companyId, CompanyTable.INSTANCE.companyId
-		).parentColumnReference(
-			UserGroupTable.INSTANCE.userGroupId,
-			UserGroupTable.INSTANCE.parentUserGroupId
+			VirtualHostTable.INSTANCE.companyId, CompanyTable.INSTANCE.companyId
+		).singleColumnReference(
+			VirtualHostTable.INSTANCE.layoutSetId,
+			LayoutSetTable.INSTANCE.layoutSetId
 		);
 	}
 
 	@Override
 	public BasePersistence<?> getBasePersistence() {
-		return _userGroupPersistence;
+		return _virtualHostPersistence;
 	}
 
 	@Override
-	public UserGroupTable getTable() {
-		return UserGroupTable.INSTANCE;
+	public VirtualHostTable getTable() {
+		return VirtualHostTable.INSTANCE;
 	}
 
 	@Reference
-	private UserGroupPersistence _userGroupPersistence;
+	private VirtualHostPersistence _virtualHostPersistence;
 
 }
